requirements.txt
flask
flask-sqlalchemy
psycopg2-binary
werkzeug
python-dotenv
docker